lms_app/requirements.txt
asn1crypto==1.5.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
contextvars==2.4
cryptography==36.0.2
dash==2.7.0
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-extensions==0.0.71
dash-html-components==2.0.0
dash-table==5.0.0
dataclasses==0.6
EditorConfig==0.12.3
Flask==2.0.3
Flask-Caching==1.10.1
idna==3.4
immutables==0.19
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
jsbeautifier==1.14.7
MarkupSafe==2.0.1
more-itertools==8.14.0
numpy==1.19.5
oscrypto==1.3.0
pandas==1.1.5
plotly==5.11.0
pycparser==2.21
pycryptodomex==3.16.0
PyJWT==2.4.0
pyOpenSSL==21.0.0
python-dateutil==2.8.2
pytz==2022.6
requests==2.27.1
six==1.16.0
snowflake-connector-python==2.7.4
tenacity==8.1.0
typing_extensions==4.1.1
urllib3==1.26.13
Werkzeug==2.0.3
zipp==3.6.0
